Updates for Puzzle Class
Completed the constructor of a new puzzle from scratch. Created helper methods to get 6 random letters and a random vowel. Completed showPuzzle method to printout the 7 letters to user. Added template for guessWord method and shuffleLetters method to be filled in later. Still need to finish the constrcutor that uses input from the user & need to find a way to make sure that the required vowel isn't already part of the 6 previous letters.